Done the Addition of Social Media Links

+++ app/Livewire/Admin/Clients/Index.php
@@ -20,14 +20,31 @@ class Index extends Component
 
     public function delete($id)
     {
-        $client = Client::find($id);
-        if($client->bookings->count() > 0) {
-            $this->dispatch('done', error: "Cannot delete client with existing bookings. Delete the bookings first.");
-            return;
-        }
-        $client->delete();
+        try {
+            $client = Client::find($id);
+            if ($client->bookings->count() > 0) {
+                throw new \Exception("Cannot delete client with existing bookings. Delete the bookings first.", 1);
 
-        $this->dispatch('done', success: "Successfully Deleted a Client");
+            }
+            if ($client->bookingRequests->count() > 0) {
+                throw new \Exception("Cannot delete client with existing booking requests. Delete the booking requests first.", 1);
+
+            }
+            if ($client->comments->count() > 0) {
+                throw new \Exception("Cannot delete client with existing comments. Delete the comments first.", 1);
+
+            }
+            if ($client->testimonial->count() > 0) {
+                throw new \Exception("Cannot delete client with an existing testimonial. Delete the testimonial first.", 1);
+
+            }
+            $client->delete();
+
+            $this->dispatch('done', success: "Successfully Deleted a Client");
+        } catch (\Throwable $th) {
+            //throw $th;
+            $this->dispatch('done', error: "Failed to Delete a Client. " . $th->getMessage());
+        }
     }
 
     public function render()
